reorder: calc p_renew_wind_onshore in one piece
rename wind.py to local_wind_onshore.py
extract renew_wind_offshore.py from wind.py
