+++ src/climatevision/generator/electricity2030/commune/electricity2030_commune.py
@@ -27,16 +27,16 @@ from ..core.energy_production.pv import (
     calc_production_local_pv_park,
     calc_production_local_pv_roof,
 )
-from ..core.energy_production.wind import (
-    calc_production_local_wind_onshore,
-    calc_renew_wind_offshore,
-)
 from ..core.energy_production.hydro import calc_production_local_hydro
+from ..core.energy_production.local_wind_onshore import (
+    calc_production_local_wind_onshore,
+)
 from ..core.energy_production.renew_hydro import calc_production_renew_hydro
 from ..core.energy_production.renew_pv_agri import calc_production_renew_pv_agri
 from ..core.energy_production.renew_pv_facade import calc_production_renew_pv_facade
 from ..core.energy_production.renew_pv_park import calc_production_renew_pv_park
 from ..core.energy_production.renew_pv_roof import calc_production_renew_pv_roof
+from ..core.energy_production.renew_wind_offshore import calc_renew_wind_offshore
 from ..core import energy_demand
 
 from .energy_production.calc_production_local_biomass_stage2 import (
@@ -189,14 +189,6 @@ def calc(
 
     p_renew_wind_offshore = calc_renew_wind_offshore(
         facts, assumptions, duration_until_target_year, d_energy=demand.total.energy
-    )
-
-    p_renew_wind_onshore = EColVars2030()
-    p_renew_wind_onshore.cost_mro_per_MWh = (
-        ass("Ass_E_P_local_wind_onshore_ratio_invest_to_power_2020")
-        * ass("Ass_E_P_local_wind_onshore_mro_per_year")
-        / fact("Fact_E_P_wind_onshore_full_load_hours")
-        * 1000
     )
 
     p_renew.CO2e_total_2021_estimated = p_renew_biomass.CO2e_total_2021_estimated
@@ -459,9 +451,34 @@ def calc(
         p_local_pv_facade_full_load_hour=p_local_pv_facade.full_load_hour,
         energy=p_renew_pv.energy * ass("Ass_E_P_renew_pv_facade_pct_of_nep_2035"),
     )
+
+    p_renew_wind_onshore = EColVars2030()
+    p_renew_wind_onshore.cost_mro_per_MWh = (
+        ass("Ass_E_P_local_wind_onshore_ratio_invest_to_power_2020")
+        * ass("Ass_E_P_local_wind_onshore_mro_per_year")
+        / fact("Fact_E_P_wind_onshore_full_load_hours")
+        * 1000
+    )
     p_renew_wind_onshore.energy = p_renew.energy * ass(
         "Ass_E_P_renew_wind_onshore_pct_of_nep_2035"
     )
+    p_renew_wind_onshore.cost_mro = (
+        p_renew_wind_onshore.energy * p_renew_wind_onshore.cost_mro_per_MWh / MILLION
+    )
+    p_renew_wind_onshore.change_energy_MWh = (
+        p_renew_wind_onshore.energy - e18.p_renew_wind_onshore.energy
+    )
+    p_renew_wind_onshore.change_cost_mro = (
+        p_renew_wind_onshore.cost_mro - e18.p_renew_wind_onshore.cost_mro
+    )
+    p_renew_wind_onshore.change_energy_pct = div(
+        p_renew_wind_onshore.change_energy_MWh, e18.p_renew_wind_onshore.energy
+    )
+    p_renew_wind_onshore.CO2e_total = 0
+    p_renew_wind_onshore.change_CO2e_t = 0
+    p_renew_wind_onshore.change_CO2e_pct = 0
+    p_renew_wind_onshore.cost_climate_saved = 0
+
     p_renew_wind_offshore.energy = p_renew.energy * ass(
         "Ass_E_P_renew_wind_offshore_pct_of_nep_2035"
     )
@@ -508,12 +525,6 @@ def calc(
         energy=p_renew_pv.energy * ass("Ass_E_P_renew_pv_agri_pct_of_nep_2035"),
     )
 
-    p_renew_wind_onshore.cost_mro = (
-        p_renew_wind_onshore.energy * p_renew_wind_onshore.cost_mro_per_MWh / MILLION
-    )
-    p_renew_wind_onshore.change_energy_MWh = (
-        p_renew_wind_onshore.energy - e18.p_renew_wind_onshore.energy
-    )
     p_renew_wind.energy = p_renew_wind_onshore.energy + p_renew_wind_offshore.energy
     p_renew_wind_offshore.cost_mro = (
         p_renew_wind_offshore.energy * p_renew_wind_offshore.cost_mro_per_MWh / MILLION
@@ -600,12 +611,6 @@ def calc(
 
     p_fossil_and_renew.change_energy_pct = div(
         p_fossil_and_renew.change_energy_MWh, e18.p_fossil_and_renew.energy
-    )
-    p_renew_wind_onshore.change_cost_mro = (
-        p_renew_wind_onshore.cost_mro - e18.p_renew_wind_onshore.cost_mro
-    )
-    p_renew_wind_onshore.change_energy_pct = div(
-        p_renew_wind_onshore.change_energy_MWh, e18.p_renew_wind_onshore.energy
     )
     p_renew_wind.CO2e_combustion_based = (
         p_renew_wind.energy * p_renew_wind.CO2e_combustion_based_per_MWh
@@ -785,9 +790,6 @@ def calc(
     p_renew_wind.change_CO2e_t = 0
     p_renew_wind.cost_climate_saved = 0
 
-    p_renew_wind_onshore.CO2e_total = 0
-
-    p_renew_wind_onshore.change_CO2e_t = 0
     p_renew_wind_offshore.change_CO2e_t = 0
     p_renew_wind_offshore.CO2e_total = 0
 
@@ -802,7 +804,6 @@ def calc(
     p_renew_pv.change_CO2e_pct = 0
 
     p_renew_wind.change_CO2e_pct = 0
-    p_renew_wind_onshore.change_CO2e_pct = 0
     p_renew_wind_offshore.change_CO2e_pct = 0
 
     p_local.change_CO2e_pct = 0
@@ -813,7 +814,6 @@ def calc(
     p_renew_pv.cost_climate_saved = 0
 
     p_renew_wind.cost_climate_saved = 0
-    p_renew_wind_onshore.cost_climate_saved = 0
     p_renew_wind_offshore.cost_climate_saved = 0
 
     p_local.cost_climate_saved = 0
